agents/cartao_credito/requirements.txt
a2a-sdk[http-server]
httpx
pydantic
uvicorn
langchain
langchain_openai
langchain-community
langchain-mcp-adapters
python-dotenv